feat: filter now allows for OR, AND and NOT
feat: filter passes now also raw filter queries

you can now pass anything to filters.
This is usefull for AutocompleteInput and AutocompleteArrayInput's filterToQuery option

as with the previous release it will deal with some simple cases more gracefully,
e.g.

{
    field: "value"
}

will check if the field is a string filter and will use a contain with some muliple cases

but if you do

{
    field: {
        equals: "value
    }
}

it will use this filter as-is