fix: prevent maximum call stack exception on recursive non-resource data structures
non-resource nested objects could lead to a infinit call stack exception,
because it would try to fetch the whole tree.

We limit the depth of non-resource objects to 1, this does not affect the tests

We later might even consider to limit it to 0 and either encourage making them resources
or to define custom views if you need nested data. We then would only fetch "primitiv" values
on an object by default and users would need to opt-in for nested data by defining them in a custom view fragment.